move file_reader and sparse_matrix to source_base
