+++ source/source_lcao/module_deepks/test/CMakeLists.txt
@@ -39,7 +39,6 @@ set(DEEPKS_UNIT_COMMON_SOURCES
     ../../../source_cell/sep.cpp
     ../../../source_cell/sep_cell.cpp
     ../../../source_pw/module_pwdft/soc.cpp
-    ../../../source_io/module_output/sparse_matrix.cpp
     ../../../source_cell/read_pseudo.cpp
     ../../../source_estate/param_update.cpp
     ../../../source_cell/cal_wfc.cpp
